Add a verbose logging option and put expensive debug output behind it by default
Logging every send/receive/request/response is very expensive for structs
due to the way they're stringified by capnproto.

Introduce the concept of verbose logging, and expose it to clients by adding
a new constructor for EventLoop which accepts LogOptions. For backwards
compatibility, the existing constructor is unchanged.

This is a behavioral change; users are not required to explicitly opt-in to
expensive logging explicitly